Fixes for original sequel adapter
There was an attempt to write sequel adapter in #1030. I took commits
from this PR. This commit fixes most of the failures that are result of
applying these commits into much newer codebase. Most of the code is
also copied from the original PR.
