+++ jsonapi-resources.gemspec
@@ -29,6 +29,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'sequel'
   spec.add_development_dependency 'sequel-rails'
   spec.add_development_dependency 'activerecord', '>= 4.1'
+  spec.add_development_dependency 'database_cleaner'
   spec.add_dependency 'activesupport', '>= 4.1'
   spec.add_dependency 'railties', '>= 4.1'
   spec.add_dependency 'concurrent-ruby'
