+++ test/support/active_record/app_config.rb
@@ -1,0 +1,13 @@
+TestApp.class_eval do
+  config.active_record.schema_format = :none
+
+   if Rails::VERSION::MAJOR >= 5
+    config.active_support.halt_callback_chains_on_return_false = false
+    config.active_record.time_zone_aware_types = [:time, :datetime]
+    config.active_record.belongs_to_required_by_default = false
+
+     if Rails::VERSION::MINOR >= 2
+      config.active_record.sqlite3.represent_boolean_as_integer = true
+    end
+  end
+end
