+++ test/support/active_record/import_schema.rb
@@ -1,0 +1,11 @@
+require 'active_record'
+
+connection = ActiveRecord::Base.connection
+
+sql = File.read(File.expand_path('../../database/dump.sql', __FILE__))
+statements = sql.split(/;$/)
+statements.pop  # the last empty statement
+
+statements.each do |statement|
+  connection.execute(statement)
+end
