+++ test/support/sequel/initialize.rb
